:watch: :bug:  Revert to using a fixed integration timestep in TOTG (#5)
PR #3 had a sneaky
change, where the integration timestep used in TOTG was change from a fixed
0.001 s (1 ms) to a proportion of the `resample_dt_` (specifically, 10%
of `resample_dt`). This change was undocumented by Moveit and just
included in the ITLP feature -____-; it is also the behavior in Noetic.

This PR reverts from the adaptive timestep to the fixed timestep, which
is what is used in Melodic TOTG.