:chart_with_upwards_trend: :tada: Add torque-stuffing to `retimeTrajectory()` (#6)
Completes SW-1869

This PR add an arg, `try_torque_stuffing`, to `retimeTrajectory()` --
when `true`, a helper will be called to add torques to the effort field
of a `RobotTrajectory.msg`. This PR **_does not_** change the actual
torque computation at all -- there is no guarantee that the torques are
accurate.